+++ apicast/src/configuration_store.lua
@@ -1,24 +1,34 @@
 local setmetatable = setmetatable
-local ipairs = ipairs
 local pairs = pairs
 local insert = table.insert
 local concat = table.concat
-local next = next
+local rawset = rawset
 
 local env = require 'resty.env'
+local lrucache = require 'resty.lrucache'
 
 local _M = {
   _VERSION = '0.1',
-  path_routing = env.enabled('APICAST_PATH_ROUTING_ENABLED')
+  path_routing = env.enabled('APICAST_PATH_ROUTING_ENABLED'),
+  cache_size = 1000
 }
 
 local mt = { __index = _M }
 
-function _M.new()
+function _M.new(cache_size)
   return setmetatable({
-    services = {},
-    hosts = {},
-    cache = {}
+    -- services hashed by id, example: {
+    --   ["16"] = service1
+    -- }
+    services = lrucache.new(cache_size or _M.cache_size),
+
+    -- hash of hosts pointing to services, example: {
+    --  ["host.example.com"] = {
+    --    { service1 },
+    --    { service2 }
+    --  }
+    cache = lrucache.new(cache_size or _M.cache_size)
+
   }, mt)
 end
 
@@ -30,8 +40,8 @@ function _M.all(self)
     return nil, 'not initialized'
   end
 
-  for _,service in pairs(all) do
-    insert(services, service.serializable or service)
+  for _,v in pairs(all.hasht) do
+    insert(services, v.serializable or v)
   end
 
   return services
@@ -44,7 +54,7 @@ function _M.find_by_id(self, service_id)
     return nil, 'not initialized'
   end
 
-  return all[service_id]
+  return all:get(service_id)
 end
 
 function _M.find_by_host(self, host)
@@ -52,72 +62,75 @@ function _M.find_by_host(self, host)
   if not cache then
     return nil, 'not initialized'
   end
-  return cache[host] or { }
+
+  return cache:get(host) or { }
 end
 
-function _M.store(self, config)
-  self.configured = true
-  local services = config.services
+local hashed_array = {
+  __index = function(t,k)
+    local v = {}
+    rawset(t,k, v)
+    return v
+  end
+}
 
-  for i = 1, #services do
-    _M.add(self, services[i])
+function _M.store(self, config, ttl)
+  self.configured = true
+
+  local services = config.services
+  local by_host = setmetatable({}, hashed_array)
+
+  local ids = {}
+
+  for i=1, #services do
+    local hosts = services[i].hosts or {}
+    local id = services[i].id
+
+    if not ids[id] then
+      ngx.log(ngx.INFO, 'added service ', id, ' configuration with hosts: ', concat(hosts, ', '))
+
+      for j=1, #hosts do
+        local h = by_host[hosts[j]]
+
+        if #(h) == 0 or _M.path_routing then
+          insert(h, services[i])
+        else
+          ngx.log(ngx.WARN, 'skipping host ', hosts[j], ' for service ', id, ' already defined by service ', h[1].id)
+        end
+      end
+
+      self.services:set(id, services[i]) -- FIXME: no ttl here, is that correct assumption?
+      ids[id] = services[i]
+    else
+      ngx.log(ngx.WARN, 'skipping service ', id, ' becasue it is a duplicate')
+    end
+  end
+
+  local cache = self.cache
+
+  for host, services_for_host in pairs(by_host) do
+    cache:set(host, services_for_host, config.ttl or ttl or _M.ttl)
   end
 
   return config
 end
 
-function _M.reset(self)
+function _M.reset(self, cache_size)
   if not self then
     return nil, 'not initialized'
   end
 
-  self.services = {}
-  self.hosts = {}
-  self.cache = {}
+  self.services = lrucache.new(cache_size or _M.cache_size)
+  self.cache = lrucache.new(cache_size or _M.cache_size)
   self.configured = false
 end
 
 function _M.add(self, service)
-  local hosts = self.hosts
-  local all = self.services
-  local cache = self.cache
-
-  if not hosts or not all then
+  if not self.services then
     return nil, 'not initialized'
   end
 
-  local id = service.id
-
-  if not service.hosts then
-    ngx.log(ngx.WARN, 'service ', id, ' is missing hosts')
-    return
-  end
-
-  for _,host in ipairs(service.hosts) do
-    local index = hosts[host] or {}
-    local exists = not _M.path_routing and next(index)
-
-    -- already exists in cache
-    local pos = index[id]
-    local c = cache[host] or {}
-
-    if pos then
-      c[pos] = service
-    else
-      insert(c, service)
-      index[id] = #c
-      all[id] = service
-    end
-
-    cache[host] = c
-    hosts[host] = index
-
-    if exists and exists ~= id then
-      ngx.log(ngx.WARN, 'host ', host, ' for service ', id, ' already defined by service ', exists)
-    end
-  end
-
-  ngx.log(ngx.INFO, 'added service ', id, ' configuration with hosts: ', concat(service.hosts, ', '))
+  return self:store({ services = { service }})
 end
 
 return _M